JSONMark/pyproject.toml
[tool.poetry]
name = "jsonmark"
version = "0.1.0"
description = ""
authors = ["Jason Yan <tailofthesun@gmail.com>"]

packages = [
    { include = "jsonmark", from = "." }
]

[tool.poetry.dependencies]
python = "^3.8"
mimesis = "^4.0.0"
click = "^7.1.2"
tqdm = "^4.46.0"
ujson = "^2.0.3"
python-rapidjson = "^1.5"
rich = "^11.0.0"
orjson = "^3.6.6"
pysimdjson = "^4.0.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
meson = "^0.54.2"

[tool.poetry.scripts]
jsonmark = "jsonmark.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
